Kinetic admin client: Changed Configuration type from class to interface and removed setter methods.  The implementation is moved to DefaultConfiguration.java class.
